search cocktail by name feature
